Separate visual and app layer in sales
