+++ src/discounts/components/SaleListPage/index.ts
@@ -1,2 +1,3 @@
 export { default } from "./SaleListPage";
 export * from "./SaleListPage";
+export * from "./filters";
