--- src/discounts/views/SaleList/messages.ts
@@ -1,37 +1,0 @@
-import { defineMessages } from "react-intl";
-
-const messages = defineMessages({
-  active: {
-    defaultMessage: "Active",
-    description: "sale status"
-  },
-  expired: {
-    defaultMessage: "Expired",
-    description: "sale status"
-  },
-  fixed: {
-    defaultMessage: "Fixed amount",
-    description: "discount type"
-  },
-  percentage: {
-    defaultMessage: "Percentage",
-    description: "discount type"
-  },
-  scheduled: {
-    defaultMessage: "Scheduled",
-    description: "sale status"
-  },
-  started: {
-    defaultMessage: "Started",
-    description: "sale start date"
-  },
-  status: {
-    defaultMessage: "Status",
-    description: "sale status"
-  },
-  type: {
-    defaultMessage: "Discount Type"
-  }
-});
-
-export default messages;
